`[proj] change `long double` to `double` for better performance
- the precision cost in exchange of speed is too high

+++ src/complex.cpp
@@ -1,11 +1,11 @@
 #include "../lib/complex.hpp"
 #include <cmath>
 
-long double Complex::squared_module() {
+double Complex::squared_module() {
     return std::pow(m_real, 2) + std::pow(m_im, 2);
 }
 
-long double Complex::module() {
+double Complex::module() {
     return std::sqrt(squared_module()); 
 }
 
@@ -15,8 +15,8 @@ Complex Complex::operator+(Complex const &c_number) {
 
 // (a + bi)^2 = (a^2 - b^2) + i(2ab)
 Complex Complex::square() {
-    long double real_part{std::pow(m_real, 2) - std::pow(m_im, 2)};
-    long double im_part{2 * m_real * m_im};
+    double real_part{std::pow(m_real, 2) - std::pow(m_im, 2)};
+    double im_part{2 * m_real * m_im};
 
     return Complex(real_part, im_part);
 }
